Added & fixed descriptions, modified stats of the EHG-22 and SHAG-1, updated HAACIcon, adjusted cost for turrets.
Weakened the armour of the Sentinel and changed all the robots butcher products in accordance to the new costs of the robots. Lowered the armour pen of the EHG-22 and made adjustments to the SHAG-1. Added descriptions to the Defender race file. Still need to do it for the Enforcer and Sentinel.